Refactor/102 (#104)
* chore: 컴포넌트 폴더, 파일 이름 컨벤션 통일

* chore: 메인logo, 검색창 분리

* chore: logo 컴포넌트화, NavLinks 분리

* fix: 글쓰기 modal이 css가 적용되지 않던 것 수정

* fix: 글쓰기 버튼이 동작하지 않던 오류 수정

* chore: 로딩스피너 수정 완료

* chore: myInfoForm 분리

* chore: userData 받아오는 과정 util함수로 분리

* chore: myInfo 전체적인 form 컴포넌트화 완료
