+++ src/components/myInfo/MyInfo.jsx
@@ -1,0 +1,7 @@
+import MyInfoForm from './myInfoForm/myInfoForm';
+
+export default function MyInfo() {
+  return (
+    <MyInfoForm />
+  );
+}
